Split playback progress ownership
